pyproject.toml
[project]
name = "smart-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "accelerate>=1.11.0",
    "fastapi[standard]>=0.118.0",
    "hf-xet>=1.1.10",
    "langchain>=1.0.0",
    "langchain-community>=0.4",
    "langchain-huggingface>=1.0.0",
    "langgraph>=0.6.10",
    "llama-index>=0.14.4",
    "pinecone>=7.3.0",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.1.1",
    "sentence-transformers>=5.1.1",
    "transformers>=4.57.0",
    "uvicorn>=0.37.0",
]
